move cli_utils tests into test_ui and update
